+++ packages/gamut-kit/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [0.5.11](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-kit@0.5.10...@codecademy/gamut-kit@0.5.11) (2021-08-03)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
+### [0.5.10](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-kit@0.5.9...@codecademy/gamut-kit@0.5.10) (2021-08-03)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
+### [0.5.9](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-kit@0.5.8...@codecademy/gamut-kit@0.5.9) (2021-08-03)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
+### [0.5.8](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-kit@0.5.7...@codecademy/gamut-kit@0.5.8) (2021-08-03)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
+### [0.5.7](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-kit@0.5.6...@codecademy/gamut-kit@0.5.7) (2021-08-03)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
 ### [0.5.6](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-kit@0.5.5...@codecademy/gamut-kit@0.5.6) (2021-08-02)
 
 **Note:** Version bump only for package @codecademy/gamut-kit